+++ play_cphb/src/Chapter_3_selection_sort.cpp
@@ -3,10 +3,10 @@
 int cphb::selection_sort_v1( std::vector<int> &v )
 {
 	int inversions = 0;
-	for ( int i = 0 ; i < v.size() ; i ++ )
+	for ( size_t i = 0 ; i < v.size() ; i ++ )
 	{
 		int minIdx = i;
-		for ( int j = i+1; j < v.size(); j ++ )
+		for ( size_t j = i+1; j < v.size(); j ++ )
 		{
 			if ( v[j] < v[minIdx] )
 				minIdx = j;
