+++ play_daily_with_one_problem/src/29_07_2019.cpp
@@ -47,7 +47,7 @@ namespace daily_interview
 				if( has_cycle( *it, visited, vertex ))
 					return true;
 			}
-			else if ( *it != parent )
+			else if ( *it != (int) parent )
 				return true;
 		}
